make error component reusable
